+++ src/torch/cuda/ops_impl_cuda.cu
@@ -225,7 +225,7 @@ void uniform_impl(Tensor& a, double mi, double ma)
     std::uniform_int_distribution<uint64_t> dist(0, std::numeric_limits<uint64_t>::max());
     uint64_t seed = dist(generator());
     int64_t max_threads = std::min<int64_t>(a.numel(), int64_t(1024) * 1024 * 1024);
-    CUDA_SWITCH_MACRO_ALL(a.device(), a.scalar_type(), max_threads, rand_float_impl, a, (float)mi, (float)ma, seed);
+    CUDA_SWITCH_MACRO_ALL(a.device(), a.scalar_type(), max_threads, rand_float_impl, a, (float)mi, (float)ma, uint32_t(seed));
 }
 
 template <typename T>
@@ -256,7 +256,7 @@ void uniform_int_impl(Tensor& a, int64_t low, int64_t high)
     std::uniform_int_distribution<uint64_t> dist(0, std::numeric_limits<uint64_t>::max());
     uint64_t seed = dist(generator());
     int64_t max_threads = std::min<int64_t>(a.numel(), int64_t(1024) * 1024 * 1024);
-    CUDA_SWITCH_MACRO_ALL(a.device(), a.scalar_type(), max_threads, rand_int_impl, a, low, high, seed);
+    CUDA_SWITCH_MACRO_ALL(a.device(), a.scalar_type(), max_threads, rand_int_impl, a, low, high, uint32_t(seed));
 }
 
 template <typename T>
